fixup! More tests
