+++ cmd/repo-updater/repos/other_external_services_test.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"net/url"
 	"reflect"
-	"sync"
 	"testing"
 
 	"github.com/kylelemons/godebug/pretty"
@@ -64,6 +63,11 @@ func TestOtherReposSyncer_syncAll(t *testing.T) {
 			Kind:   "OTHER",
 			Config: `{"repos": [""]}`,
 		},
+		"invalid-json": {
+			ID:     2,
+			Kind:   "OTHER",
+			Config: `{`,
+		},
 	}
 
 	repos := map[string]*api.Repo{
@@ -115,6 +119,30 @@ func TestOtherReposSyncer_syncAll(t *testing.T) {
 			},
 		},
 		{
+			name:   "external service listing error",
+			svcs:   []*api.ExternalService{},
+			before: []*api.Repo{},
+			after:  []*api.Repo{},
+			err:    errors.New(`no external services of kind "OTHER"`),
+		},
+		{
+			name:   "invalid JSON in exernal service config",
+			svcs:   []*api.ExternalService{svcs["invalid-json"]},
+			before: []*api.Repo{},
+			after:  []*api.Repo{},
+			results: SyncResults{
+				{
+					Service: svcs["invalid-json"],
+					Errors: SyncErrors{
+						{
+							Service: svcs["invalid-json"],
+							Err:     "config error: failed to parse JSON: [CloseBraceExpected]",
+						},
+					},
+				},
+			},
+		},
+		{
 			name:   "invalid external service configs return an error",
 			svcs:   []*api.ExternalService{svcs["bad"]},
 			before: []*api.Repo{},
@@ -160,113 +188,4 @@ func TestOtherReposSyncer_syncAll(t *testing.T) {
 			}
 		})
 	}
-}
-
-// FakeInternalAPI implements the InternalAPI interface with the given in memory data.
-// It's safe for concurrent use.
-type FakeInternalAPI struct {
-	mu     sync.RWMutex
-	svcs   map[string][]*api.ExternalService
-	repos  map[api.RepoName]*api.Repo
-	repoID api.RepoID
-}
-
-// NewFakeInternalAPI returns a new FakeInternalAPI initialised with the given data.
-func NewFakeInternalAPI(svcs []*api.ExternalService, repos []*api.Repo) *FakeInternalAPI {
-	fa := FakeInternalAPI{
-		svcs:  map[string][]*api.ExternalService{},
-		repos: map[api.RepoName]*api.Repo{},
-	}
-
-	for _, svc := range svcs {
-		fa.svcs[svc.Kind] = append(fa.svcs[svc.Kind], svc)
-	}
-
-	for _, repo := range repos {
-		fa.repos[repo.Name] = repo
-	}
-
-	return &fa
-}
-
-// ExternalServicesList lists external services of the given Kind. A non-existent kind
-// will result in an error being returned.
-func (a *FakeInternalAPI) ExternalServicesList(
-	_ context.Context,
-	req api.ExternalServicesListRequest,
-) ([]*api.ExternalService, error) {
-
-	a.mu.RLock()
-	defer a.mu.RUnlock()
-
-	svcs, ok := a.svcs[req.Kind]
-	if !ok {
-		return nil, fmt.Errorf("no external services of kind %q", req.Kind)
-	}
-
-	return svcs, nil
-}
-
-// ReposList returns the list of all repos in the API.
-func (a *FakeInternalAPI) ReposList() []*api.Repo {
-	a.mu.RLock()
-	defer a.mu.RUnlock()
-
-	repos := make([]*api.Repo, 0, len(a.repos))
-	for _, repo := range a.repos {
-		repos = append(repos, repo)
-	}
-
-	return repos
-}
-
-// RepoCreateIfNotExists creates the given repo if it doesn't exist. Repos with
-// with an empty name are invalid and will result in an error to be returned.
-func (a *FakeInternalAPI) ReposCreateIfNotExists(
-	ctx context.Context,
-	req api.RepoCreateOrUpdateRequest,
-) (*api.Repo, error) {
-
-	if req.RepoName == "" {
-		return nil, errors.New("invalid empty repo name")
-	}
-
-	a.mu.Lock()
-	defer a.mu.Unlock()
-
-	repo, ok := a.repos[req.RepoName]
-	if !ok {
-		a.repoID++
-		repo = &api.Repo{
-			ID:           a.repoID,
-			Name:         req.RepoName,
-			Enabled:      req.Enabled,
-			ExternalRepo: req.ExternalRepo,
-		}
-		a.repos[req.RepoName] = repo
-	}
-
-	return repo, nil
-}
-
-// ReposUpdateMetdata updates the metadata of repo with the given name.
-// Non-existent repos return an error.
-func (a *FakeInternalAPI) ReposUpdateMetadata(
-	ctx context.Context,
-	repoName api.RepoName,
-	description string,
-	fork, archived bool,
-) error {
-	a.mu.Lock()
-	defer a.mu.Unlock()
-
-	_, ok := a.repos[repoName]
-	if !ok {
-		return fmt.Errorf("repo %q not found", repoName)
-	}
-
-	// Fake success, no updates needed because the returned types (api.Repo)
-	// don't include these fields.
-
-	return nil
 }
