+++ cmd/repo-updater/repoupdater/server_test.go
@@ -5,6 +5,7 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
+	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"reflect"
@@ -14,8 +15,50 @@ import (
 	"github.com/sourcegraph/sourcegraph/cmd/repo-updater/repos"
 	"github.com/sourcegraph/sourcegraph/pkg/api"
 	"github.com/sourcegraph/sourcegraph/pkg/extsvc/github"
+	"github.com/sourcegraph/sourcegraph/pkg/repoupdater"
 	"github.com/sourcegraph/sourcegraph/pkg/repoupdater/protocol"
 )
+
+func TestServer_handleExternalServiceSync(t *testing.T) {
+	for _, tc := range []struct {
+		name string
+		svc  *api.ExternalService
+		err  string
+	}{
+		{
+			name: "bad kind",
+			svc:  &api.ExternalService{},
+			err:  "empty external service kind\n",
+		},
+		{
+			name: "bad service config",
+			svc: &api.ExternalService{
+				DisplayName: "Other",
+				Kind:        "OTHER",
+				Config:      "{",
+			},
+			err: "external-service=\"Other\": config error: failed to parse JSON: [CloseBraceExpected]; \n",
+		},
+	} {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			t.Parallel()
+
+			fa := repos.NewFakeInternalAPI([]*api.ExternalService{tc.svc}, nil)
+			s := Server{OtherReposSyncer: repos.NewOtherReposSyncer(fa, nil)}
+			ts := httptest.NewServer(s.Handler())
+			defer ts.Close()
+
+			cli := repoupdater.Client{URL: ts.URL, HTTPClient: http.DefaultClient}
+			ctx := context.Background()
+
+			err := cli.SyncExternalService(ctx, *tc.svc)
+			if have, want := fmt.Sprint(err), tc.err; have != want {
+				t.Errorf("\nhave: %s\nwant: %s", have, want)
+			}
+		})
+	}
+}
 
 func TestServer_handleRepoLookup(t *testing.T) {
 	s := &Server{OtherReposSyncer: repos.NewOtherReposSyncer(api.InternalClient, nil)}
